Refactor code to library
